DIGG-482: more refactor
